+++ src/main/kotlin/com/toasttab/pulseman/entities/ProjectSettingsV3.kt
@@ -17,6 +17,8 @@ package com.toasttab.pulseman.entities
 
 data class ProjectSettingsV3(
     val configVersion: String,
+    val gradleScript: String?,
+    val javaHome: String?,
     val newJarFormatUsed: Boolean?, // This is a nullable field to support older versions of the project settings
     val tabs: List<TabValuesV3>
 ) : ProjectSettings {