+++ src/main/kotlin/com/toasttab/pulseman/state/protocol/protobuf/SendProtobufMessage.kt
@@ -87,7 +87,7 @@ class SendProtobufMessage(
     private val textArea =
         RSyntaxTextArea.textArea(
             initialSettings?.protobufSettings?.code ?: "",
-            SyntaxConstants.SYNTAX_STYLE_JAVA,
+            SyntaxConstants.SYNTAX_STYLE_KOTLIN,
             onChange
         )